+++ Tests/BugsnagUserTest.m
@@ -11,9 +11,14 @@
 #import "BugsnagUser.h"
 #import "BugsnagEvent.h"
 
+@interface BugsnagEvent ()
+- (instancetype)initWithKSReport:(NSDictionary *)report;
+@end
+
 @interface BugsnagUser ()
 - (instancetype)initWithDictionary:(NSDictionary *)dict;
 - (NSDictionary *)toJson;
+- (instancetype)initWithUserId:(NSString *)userId name:(NSString *)name emailAddress:(NSString *)emailAddress;
 @end
 
 @interface BugsnagUserTest : XCTestCase