+++ CHANGELOG.md
@@ -26,6 +26,9 @@ Bugsnag Notifiers on other platforms.
 * Hide additional methods from public API
   [#552](https://github.com/bugsnag/bugsnag-cocoa/pull/552)
 
+* Move `codeBundleId` from configuration to non-public client property
+  [#548](https://github.com/bugsnag/bugsnag-cocoa/pull/548)
+
 * Add structured app/device fields to `BugsnagSession`
   [#546](https://github.com/bugsnag/bugsnag-cocoa/pull/546)